+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.0.0-beta.3
 - Bump to Geode v5.3.0
 - Add syzzi's CBF mod support
+- Remove image indicator (I didn't rly like it)
 
 # 1.0.0-beta.2
 - Bump to Geode v5.1.0